api/requirements.txt
Django~=5.0.2
djangorestframework~=3.14.0
python-dotenv~=1.0.1
pip~=22.0.2
pipenv~=2023.12.1
pytz~=2024.1
asgiref~=3.7.2
sqlparse~=0.4.4
typing_extensions~=4.9.0
filelock~=3.13.1
certifi~=2024.2.2
setuptools~=69.0.3
virtualenv~=20.25.0
distlib~=0.3.8
platformdirs~=4.2.0
mysqlclient~=2.2.4
uwsgi~=2.0.24
dj-database-url~=2.1.0
django-cors-headers~=4.3.1
djangorestframework-simplejwt~=5.3.0